refactor(prompt): extract question draft logic and use Set lookups
